Working Hitboxes, Hurtboxes, and Pushboxes
Lots of major changes have occured to the overall archetecture of the project in this update.
For now, a lot of what used to be exposed in the inspector is now hidden (atleast until runtime) and many of the variables that were set in the inspector are now set in the code instead. If deemed neccesary, this can be brough back in the future.

Hurtboxes, Hitboxes and PhysicsBody are now children of CustomCollider2D. Making them children of the same object allows us to keep them in the same list, although we might not want this in the future.

The gizmos that showed hitboxes before have all been removed, as we moved away from monobehaviour in some cases. Moving away from monobehaviour will make it harder to set values, especially at runtime, but allow us to more easily create new instances of our classes in code without having to instantiate new gameobjects to attach them to.