Slight fixes and value tweaks
I had to leave when I committed yesterday so changes will be noted here

-A second camera is used to make a minimap, this is very simple although
it looks good
-Consider going even more simple by generating a texture based on
blocked vs empty tiles

-MapGenerator now uses the spelunky algorithm for making a path through
the map. This is a great start and will be easy to modify going forward.
-Maps now take in a MapData scriptable object when being generated -
this hold information about the maps themselves and what kind of values
they have, such as gravity
-There still remains the funciton, GenerateMap2 which uses noise to make
the entire map, its interesting but would require more work to make it
playable

MapChunks/Rooms/RoomData
-MapChunks were replaced with Rooms, and then Rooms were replaced with
RoomData, which all pretty much did the same thing but as I changed a
few things here and there it was easier to make a new class than to
modify the existing
-There are now far less types of room orientation, as thats how the
spelunky generator works, this can be expanded upon in the future

Doors and Start Tiles
-Previously, doors and start tiles were set by picking the first open
tile in a room starting in the bottom corner
-Now, they look through the room for all legal placements and then pick
one at random