Post Processing
You can now set spaces in the room editor where small or large enemies
spawn

After the map is made, the tiles are checked for enemy spawn spaces, the
enemy is spawned and then the tile is set to empty.

You can also set a new tiletype called Obstacle1, which can be either
empty, block, or oneway.

Some of these tiles have been added to the maps, and random placement
has been taken out for now.