Minor fixes
This is a follow up update to the Biosamples and Biomes update.

Fixes issue with completedWorlds array in Levelmanager becoming out of
bounds. This was happening because as new worlds were added, new nodes
for completed worlds were not. I dont really like this system as is but
am not going to bother fixing it right now.