World Chests and Auto Pickup
With this update, the loot tables for chests has been updated for each
world, and each world now has its own chest style.

Certain items will be automatically picked up when a player makes
contact with them. This flag can be set from an items prototype.

When any item is picked up, a floating text will display the name of the
item with the appropriate color determined by rarity.